+++ R/dcc.R
@@ -57,7 +57,6 @@
         # quasi-likelihood
         solution <- NULL
         pmatrix <- copy(object$parmatrix)
-        all_pars <- c(as.vector(sapply(object$univariate, function(x) x$parmatrix[estimate == 1]$value)))
         new_spec <- copy(object)
         new_spec$parmatrix <- copy(pmatrix)
         mmatrix <- .joint_parameter_matrix(new_spec)
@@ -102,7 +101,7 @@
             pmatrix[parameter == "shape", estimate := 1]
             pmatrix[estimate == 1]$value <- pars
             # hessian/scores
-            all_pars <- c(as.vector(sapply(object$univariate, function(x) x$parmatrix[estimate == 1]$value)), pmatrix[estimate == 1]$value)
+            all_pars <- c(unlist(lapply(object$univariate, function(x) x$parmatrix[estimate == 1]$value), use.names = FALSE), pmatrix[estimate == 1]$value)
             new_spec <- copy(object)
             new_spec$parmatrix <- copy(pmatrix)
             mmatrix <- .joint_parameter_matrix(new_spec)
@@ -162,7 +161,7 @@
                 pmatrix <- copy(object$parmatrix)
                 pmatrix[estimate == 1]$value <- pars
                 # hessian/scores
-                all_pars <- c(as.vector(sapply(object$univariate, function(x) x$parmatrix[estimate == 1]$value)), pmatrix[estimate == 1]$value)
+                all_pars <- c(unlist(lapply(object$univariate, function(x) x$parmatrix[estimate == 1]$value), use.names = FALSE), pmatrix[estimate == 1]$value)
                 new_spec <- copy(object)
                 new_spec$parmatrix <- copy(pmatrix)
                 mmatrix <- .joint_parameter_matrix(new_spec)
@@ -259,7 +258,7 @@
         dcc_nll <- W[["dcc_nll"]]
         model_nll <- solution$nll
         # hessian and scores
-        all_pars <- c(as.vector(sapply(object$univariate, function(x) x$parmatrix[estimate == 1]$value)), pmatrix[estimate == 1]$value)
+        all_pars <- c(unlist(lapply(object$univariate, function(x) x$parmatrix[estimate == 1]$value), use.names = FALSE), pmatrix[estimate == 1]$value)
         new_spec <- copy(object)
         new_spec$parmatrix <- copy(pmatrix)
         mmatrix <- .joint_parameter_matrix(new_spec)